Fixed Win32 prototype of Sleep().
git-svn-id: file:///home/tjaden/{repo}/redstone/trunk@429 ec8ac367-a008-0410-9901-8f3136df4969

+++ src/yield.c
@@ -27,7 +27,7 @@ void yield (void)
     /* This seems to give better results under Windows that Sleep(0)
      * (which is what yield_timeslice() uses).  */
 
-    void __stdcall Sleep (int);
+    void __stdcall Sleep (unsigned long);
     Sleep (5);
 
 #else
